Added command to create db triggers. Code file structure improved
